feat: move pycorr-related compression readers to a legacy file to remove pycorr dependency in tests
