+++ pyproject.toml
@@ -71,7 +71,6 @@ testing = [ # Testing dependencies, with minimal install for type hints
     "sunbird[inference] @ git+https://github.com/florpi/sunbird",
     # not tested, but dependency of some imports - minimal install
     "lsstypes @ git+https://github.com/adematti/lsstypes",
-    "pycorr @ git+https://github.com/cosmodesi/pycorr", 
 ]
 
 [project.urls]
